bcftbx/cli: update 'prep_samplesheet.py' and 'split_fastq.py' to use platforms.illumina code.
